worked on motor

+++ Node 1/ByggernGRP21/drivers/CAN.h
@@ -22,7 +22,7 @@
 
 
 typedef struct{
-	unsigned int ID;
+	uint8_t ID;
 	unsigned char data[8];
 	uint8_t length;
 } CANmessage;
